Restores a nullable-String concat guard even after valueOf() calls.
JSCompiler's type checker knows that nullable string + nullable string
might result in a number value, since null + null = 0 in JS. Because of
this J2CL specially augments some some string concatenation expressions
with a ("" +) to prevent that possible outcome and preserve Java semantics.

But we also specially treat String.valueOf() as if it returns a non-null
String (even though the function we output for String.valueOf() returns
a nullable string). This meant that certain expressions would not be
augmented with the ("" +) guard and would be caught by the JSCompiler
type checker.

I've removed the special case that was causing String.valueOf() to be
treated as if its return value was non-null. This special handling was only
safe back when our String.impl.java.js implementation was hand-rolled a
with a type signature that was non-null. But now that we create it from
String.java the special case is no longer safe.

PiperOrigin-RevId: 134747292