Simplify the implementation of intersection types.
This patch aims to model intersection types in a style analogous to Java type erasure. Java (through type erasure) models the intersection type as equivalent to the first type in the intersection and inserts casts whenever it is assigned to or dereferenced as other type in the intersection.

E.g.

class A<T extends Comparable<T> & CharSequence> {
  CharSequence g(T t) {
    t.compareTo(null);  // Comparable method
    t.length();  // CharSequence method
    return t;
  }
}

gets translated as

class A<T extends Comparable<T>> {
  CharSequence g(T t) {
    t.compareTo(null);
    ((CharSequence) t).length();
    return ((CharSequence) t);
  }
}

PiperOrigin-RevId: 136642378