Adds an extensive bridge method regression suite.
Adds 12000 bridge method test cases in a new /regression/ directory.

About 3000 of them are initially disabled since we handle them wrong.

The commented out test cases should be re-enabled as bridge method
construction is rewritten and fixed.

Test cases are variations of the following class hierarchy:

Constructs variations of this class hierarchy:
  class A {}
  interface BI3 {}
  interface BI2 extends BI3 {}
  interface BI1 extends BI3 {}
  class B extends A implements BI1, BI2 {}
  interface CI2 extends CI3 {}
  interface CI1 extends CI3 {}
  class C extends B implements CI1, CI2 {}

Variations are generated on the following axes:
  class/interface existence
  hasClassMethod
  hasInterfaceMethod
  hasDefaultMethod
  isAbstract
  isJsType
  typeArgument presence
  typeArgument bound
  typeParameter presence
  typeParameter bound

PiperOrigin-RevId: 136875180