Remove unnecessary set -e
Per go/be;

The genrule command is executed in a bash shell (from BuildHelpers). The bash shell is configured to fail when a command fails (set -e), a command in a pipeline fails (set -o pipefail), or an undefined variable is expanded (set -u). Genrules should not access the network (except to create connections between processes running within the same genrule on the same machine), though this is not currently enforced in all cases.

PiperOrigin-RevId: 138606466