Fixed a bug in construction of certain TypeDescriptors.
Creation of TypeDescriptors referencing specializations of generic local classes
was failing to propagate the type variables inherited from the declaration scope.

This was resulting in JsDoc references to the type having too few generic
specializations compared to the class definition and also sometimes referencing
the wrong type parameters (referencing ones in the enclosing class instead of
ones from the declaration scope).

PiperOrigin-RevId: 138672373