Centralizes anonymous class construction in CompilationUnitBuilder.
Previously the transformation of anonymous classes would be done in two steps, first CUB would an AnonymousType (a subclass of Type, now gone), stash some data related to captures and super constructor invocation there and then a second fixup pass, FixAnonymousClassConstructor (now also gone),  would use that information finalize the transformation.

There was not really a need to split into two passes as the constructor for the anonymous type can be created by CUB removing the need for stashing and fixups.

This cl slightly increases the complexity of CUB momentarily, a follow up cl that will simplify CUB in a significant way.

PiperOrigin-RevId: 139138979