Remove unnecessary extra lambda method and remove some duplication in forwarding method construction.
Lambdas are currently implemented with 2 methods, a private one containing the actual code written by the user and a small forwarding stub that redirects interface methods to the implementation. This scheme is not necessary in J2cl and this CL removes the extra method by synthesizing the lambda code in the implementation of the interface method.

PiperOrigin-RevId: 141491076