Remove type variable references inside anonymous (arrow) functions.
Closure template variables declared in the class or the enclosing method are not available inside the body of anonymous (arrow) functions.

This CL expands the current fixup, which removed removed references to template variables declared by methods in the method body because of closure restrictions, to handle anonymous functions as well.

PiperOrigin-RevId: 141499068