Splits TypeDeclarationDescriptor out of TypeDescriptor.
Tracking these two concepts in separate classes makes it possible to use to Java type checker to verify that all synthetic conversions between them are known and properly handled (which for now means making sure that we have created all possible TypeDescriptors first, so that when synthesizing a TypeDescriptor from a TypeDeclarationDescriptor our interner will redirect the created instance to the correct one).

Previously we modeled both concepts in just the TypeDescriptor class but this was (in rare circumstances) vulnerable to intern key collision (because of blurry lines between type parameters and type arguments) leading to holding a reference to the wrong instance of a usage type descriptor.

PiperOrigin-RevId: 146503640