Add more detailed readable for string conversions.
The implementation of string conversions (InsertStringConversions.java:67) assumes that some operations would not need boxing and are skipped however by the time the conversions happen boxing has already occurred.

The description in the JLS regarding string conversions (JLS 5.1.11) describes the algorithm in terms of creating boxes for the primitives (although not exactly the same boxes that boxing uses).

PiperOrigin-RevId: 146934937