Add precondition check that was implicitly removed by cl/147862498.
Autovalue used to enforce nullability when construction the value object but allowed the corresponding setters in builders to receive null. A forecoming change in autovalue adds the restriction to the setters in the builders as well. cl/147862498 made those fields nullable (removing the checks altogether) and this one adds them back explicitly in the builder.

PiperOrigin-RevId: 147865774