Fix for intersection types bug b/35625063.
NormalizeIntersectionTypes converts intersection casts like (A & B) into multiple casts to basic types such as (A)(B) and also inserts specific casts to a member of the intersection when accessing members or doing operations that turn an intersection cast into a single cast.

NormalizeIntersectionTypes when doing this second rewriting assumed that the destination type was a base type (class, enum or interface) which is not always true.

In most cases we would not need Java casts to go from an intersection cast to a member cast; a JsDoc annotation would suffice. Most of these casts are redundant as the erasure pass will insert them. There is only one case in which a needed cast will not be inserted by the erasure pass and that is when specializing from <T extends A> to <T extends A & B> because function with T as parameter both have the same signature and will not result in a bridge.

PiperOrigin-RevId: 148242183