Fixing a bug in BridgeMethodsCreator.
Its strategy can be described as follows:
- make a list of potential bridge method signatures
- try to find a concrete dispatch target for each bridge method
- failing that, try to dispatch it "backwards" as an accidental override.

It contained a bug where it was giving default methods too high a priority as a dispatch target. It would sometimes find and use concrete dispatch target that happened to be a default method even though the "fallback" accidental override that was available should have been higher priority (since it would result in targetting a regular concrete method on the prototype chain).

PiperOrigin-RevId: 148791979