Adds a repro case for an abstract method stub versus bridge method issue in AutoValue generated code.
When required abstract methods (abstract either because they're created abstract in a parent class or because they are required by an interface)
with a particular name are provided in both @JsMethod and non@JsMethod forms both the JsBridgeMethodsCreator and UnimplementedMethodsCreator passes might need to do some work (creating both m_foo__() and foo() forms of the method) but importantly ONLY ONE of them should create the missing methods.

Both of them look at the circumstances and then decide whether it is a situation they are responsible for or not.

In the past if abstract class 'B' defined a non@JsMethod abstract method but that same abstract method was already required by some implemented interface but in @JsMethod form, neither JsBridgeMethodCreator nor UnimplementedMethodsCreator would fill in the gap.

Rolled back cl/149033705 fixed this by making UnimplementedMethodsCreator aware of foo() forms of abstract methods but this created collisions when BridgeMethodsCreator and JsBridgeMethodsCreator would create colliding bridges in abstract classes. The CL heavy-handedly disabled all such bridge method creation by those passes but this was not good enough since AutoValue constructs some class scenarios where it would result in neither bridges or abstract stubs being created:

  interface P {
    String getKey();
  }

  @JsType
  @SuppressWarnings("ClassCanBeStatic")
  abstract class Q implements P {
    @Override
    public abstract String getKey();
  }

  @JsType
  abstract class R extends Q {
    @Override
    public String getKey() {
      return null;
    }
  }

  class S extends R {}

This CL adds the repro case and it will be addressed in a followup CL.

PiperOrigin-RevId: 150219775