Update jsinterop restriction checker to enforce restrictions on JsConstructor subclasses.
There are subtle changes in the spec that should be reflected in the design doc:
- subclasses of JsConstructor classes need to have a JsConstructor (that delegates to super). There is no longer the concept of primary contructors.
- anonymous inner classes of a JsConstructor supertype have are implicitly JsConstructor classes.

The condition of being a primary constructor (the only one delegating to super) requires examination of the code which is not accessible from classes compiled separately. The introduction of this concept was motivated by GWT so as to not require to export subclasses of native classes. Such requirement does not exist in J2CL and this new semantic is simpler to understand and communicate, is checkeable and it is simpler to implement.

PiperOrigin-RevId: 150258492