Fixes another missing abstract method warning.
This is a reroll of rolledback cl/149033705 with changes to correct its error.

For ease of review i've diffbased on cl/150224024 so you can see just the new and different parts.

The error was that the original CL stopped creating JsBridge methods in abstract classes when what it should have done was stop creating JsBridge methods for abstract methods. The failure to create JsBridge methods for concrete methods defined in abstract classes simply resulted in some missing bridges.

See cl/150224024 for explanatory comments on the hidden diffbased portions.

PiperOrigin-RevId: 150697227