Automated g4 rollback of changelist 151173662.
*** Reason for rollback ***

I'm rolling this back because it introduces a huge 300k code size regression. Diffing the PMF files, none of the input JS is changed, but I believe it alters the dependency graph in a way that makes PMF enhancer change the module dependencies. All of the changes between PMFs were in what modules depended on what other synthetic modules.

This regression is blocking releases, so I need to roll it back, and then roll it forward when the size regression is tracked down.

*** Original change description ***

Rollforward of cl/150859899 again with fixes to xplat_async APT processor and xplat_j2cl_test_suite

Because the j2cl_library rule now automatically picks up .js files generated by APT processors, xplat_async's macros do not need to separately extract the JS. On the other hand,
the xplat_j2cl_test_suite rule has to extract the JS and bundle it as a zip for input to
closure_test_suite, therefore, output of the XplatTestProcessor is cloaked from the j2cl_library rule by appending .testsuite as a...

***

PiperOrigin-RevId: 151453994