Reroll: Insert a cast around method calls that return a generic.
This works around an issue where JSCompiler type inference can lead to a
type mismatch due to type tightening.

This warning fix had been removed at a point because it was creating
violations of AJD strict deps checking. But AJD now allows J2CL to
violate strict deps. So this is no longer a concern.

PiperOrigin-RevId: 151728567