Use supported web test infrastructure to provision browsers.
Web test infrastructure maintains the proper dependencies and information required for browsers to be provisioned as well as browser information and APIs to be correct; the logic that falls back to inferring partial browser information from the obsolete system property is being turned down.

Note that web_test has been used (rather than jsunit_test_suite) to preserve the names of the targets and, thus, the target history as well as prevent certain behavior currently implemented in jsunit_test_suite from triggering. If you would prefer to use the jsunit_test_suite macro, please produce a changelist to do so.

Tested:
    TAP --sample for global presubmit queue
    http://test/OCL:151765297:BASE:151743249:1490916139607:a0475522
PiperOrigin-RevId: 152319791