Make sure that FixTypeVariablesInMethods only leaves references to type variables defined by the class.
It seems that some of our passes are not very precise and may leak type variables that are not defined in the method nor in the enclosing class (even though they might be rendered with the same name).

Ideally the verification pass should check the invariant that only type variables that are defined in the enclosing scope appear in the method bodies; and the passes that violate the invariant need to be fixed.

PiperOrigin-RevId: 153529564