Add module_graph_root to pinto_processed_module_set and its wrapper macros.
Before this change, PmfFileEnhancer inferred that the first module in deps is the root of the module graph, and added a dependency between every module and that module. This is very confusing to users and has caused many subtle bugs that were difficult to catch.

See the LSC doc (http://go/static-module-pinto-processed-module-set) for more details.

Tested:
    TAP --sample for global presubmit queue
    http://test/OCL:155656337:BASE:155648011:1494442168405:b52e8189
PiperOrigin-RevId: 155685012