Remove unnecessary runs of PureFunctionIdentifier in J2ClClinitPrunerPass.
Only one of the optimizations may change the side-effect of the functions which is EmptyClinitPruner. Others only removes the redundant ones so should not have any impact on side-effect state.

Per cl/159465374, PureFunctionIdentier is taking 4.0% percent of total run-time of kix, so I'm guessing this should significantly help with that.

No code size changes in kix or j2cl tests.

PiperOrigin-RevId: 164914270