Optimize more J2CL clinit patterns.
When clinit X just exists to call another clinit Y, we can replace all clinit X reference with clinit Y which will result in clinit X to be unused and get pruned. And that is exactly what this CL does.

This could be quite common if one of the very top level classes have a clinit that cannot be optimized. In such case all subclasses (recursively) need to keep a clinit just to call the other one.

PiperOrigin-RevId: 169185339