Stops emitting stubs in boxed primitives and JsFunction implementations.
J2CL was emitting stub methods for devirtualized functions defined in boxed primitives and JsFunction implementations. The reason was to make JsCompiler happy about the parent contracts that weren't fulfilled. However JsCompiler cannot always identify that they are not used nor have a concept of if a type instantiated or not. Hence they were not pruned.

This CL marks such types as abstract and no longer emit the stubs. We still might later emit abstract methods to fulfill the contracts but those are guaranteed to be removed.

To make this work, I needed to update UnimplementedMethodsCreator to handle removed methods (it wasn't looking at Type earlier).

PiperOrigin-RevId: 174938235