Ensure streams that encapsulate a closeable resource are closed
This cleanup is being performed in preparation for enabling a
compiler error.

From the javadoc for Files.list:

The returned stream encapsulates a DirectoryStream. If timely disposal
of file system resources is required, the try-with-resources construct
should be used to ensure that the stream's close method is invoked
after the stream operations are completed.

More information:
https://docs.google.com/document/d/1ReWb1eSJxAhBZg0P9FqxHUtTW0GemZJVXUhEPaFywuU/edit?usp=sharing

Tested:
    tap_presubmit: http://test/OCL:180815228:BASE:180939806:1515205155738:3e3be203
    Some tests failed; test failures are believed to be unrelated to this CL
PiperOrigin-RevId: 181236670