Handle intersection types correctly in the backend.
There are currently two passes that handle intersection types (NormalizeIntersectionTypes and InsertErasureTypeSafetyCasts) but those do not completely remove intersection types and some might percolate to the generation stage.

As a first step to clean up, this cl implements the handling in the generation stage. A follow up cl will simplify the rewriting.

PiperOrigin-RevId: 181505229