Makes all new lines consistent with unix newline
This also removes unnecessary apache-lang3 dep.

We were not fully consistent with the use of new lines which
could cause several issues around source maps and RTA. Since
this is generated code we can choose to be consistent on a single
platform and most (if not all) development tools will handle
that in all platform if somebody cares to take a look.

PiperOrigin-RevId: 217374862