Disallow Native JsEnum type literal.
Since native JsEnums are not boxed, getClass() on a native JsEnum value is that of the underlying type. The underlying type is not always known to J2CL (e.g. when the native JsEnum does not have custom value), making it impossible for NativeJsEnum.class be consistent with NativeJsEnum.CONSTANT.getClass().

PiperOrigin-RevId: 218222493