Insert boxing operations when types are specialized.
The current ContextRewriter does not handle conversions when type variables are specialized. Before the introduction of JsEnums, this was only needed for erasure casts on returns and field accesses and those cases were special handled in that pass. With the introduction of JsEnum, unboxing needs to happen when types are specialized in returns and field accesses (similar to erasure casts). In addition boxing might need to happen when assigning a parameter or a field whose type was specialized to JsEnum.

In this CL we expand the ContextRewriter to provide the declared type in addition to the inferred type for assignment and method invocation contexts to handle boxing on assignments in all situations. And we add special casing for unboxing on specialization similar to the one in InsertErasureTypeSafetyCasts.

PiperOrigin-RevId: 219714036