Avoid NPE during restriction checking.
Passes in the transpiler make assumptions that the classes being processed satisfy jsinterop restrictions. That fact is not true while performing the restriction checking and code executed by the checker should not make those assumptions.

This cl fixes one such instance is one in which a JsFunction implementation is declared abstract and thus does not have a JsFunction method.

PiperOrigin-RevId: 219759195