Rewrite calls to method from java.lang.Enum when called on a JsEnum instance.
Since JsEnum classes do not really extend Enum, method calls to Enum methods need to be rewritten to keep the AST in a sane state. Methods that are overridden by Enum are dispatched to their overridden method (e.g. Comparable.compareTo or Object.toString) and methods that are specific (like ordinal()) are rewritten as if they were declared by the specific JsEnum instance since they will be handled specially (akin to Enum.values and Enum.valueOf for normal enums, which do really exist in the source but are assumed to exist after the class is compiled).

PiperOrigin-RevId: 220267056