Fix console commands code blocks.
The convention for console commands code blocks are different in different markdown implementations.

Github uses the convention of indenting by 6 spaces and not using backticks, whereas internal tooling uses an indented and backticked code block. Neither are rendered properly in the other implementation.

As a compromise I found that a backticked block with indented contents renders reasonably in both.

PiperOrigin-RevId: 221002078