Generalize ConversionContextVisitor to handle instance member qualifiers as assignment contexts.
This simplifies passes like InsertErasureTypeSafetyCasts and NormalizeIntersectionTypes. Additionally handling all assignments consistently is clearly beneficial.

NormalizeIntersectionTypes was handling two separate aspects of intersection types:
1. rewriting intersection casts to plain casts, e.g. (A&B) => (B)(A)
2. making sure the intersection was cast to the right component before a method call or field access, e.g. suppose the following code:

  interface A {}
  interface B {  void m(); }
  class X<T extends A&B> {
    void method(T t) {
      t.m();
    }
  }

the signature of X.method is method(A), and Java code can only statically guarantee that it is called with an instance of A because you can always subvert generics with raw casts, e.g new X().method(new A() {}) is valid Java code.

Number 2. is basically the same as the insertion of an erasure casts now that the casts are inserted like in javac, i.e. to the declared type of the enclosing type of the member. The only modification needed in the erasure pass was to change the assignability check to use the raw type which is the only type that is statically guaranteed.

PiperOrigin-RevId: 224659910