Fix bad expansion of unary expressions.
Since J2CL backend (ExpressionTranspiler) does not insert parenthesis to preserve correct precedence, but instead relies the parenthesis being explicitly represented in the AST (see b/67753876) extra care needs to be taken when performing transformations.

In this case expressions like ++a might get expanded to a = a + 1 where the assignment operator = is of much lower precedence than the decrement. This might be problematic in expressions like ++a < 4  where the lack of parenthesis after expansion to a = a + 1 < 4 changes its meaning.

In a sane world, parenthesis would not be explicitly represented in the IR but just computed upon code generation. The actual order of operations is dictated by the structure of the tree, making them unnecessary. This is not as grim as it seems since almost all transformations insert calls or constructs of the highest precedence.

PiperOrigin-RevId: 229261436