Avoid running java_common.compile when there are no java srcs.
Even there are no srcs, we create dummy srcs pass that to java_common.compile so it doesn't choke (b/110982833). This creates unnecessary workload in the build graph (not sure how much) as it needs to trigger javac actions.

I'm adding a simple work around so this doesn't happen anymore.

PiperOrigin-RevId: 230839048