Support the workflow under Windows. #9
Use hardcoded '/' as the file separator in getRelativePath and getAbsolutePath.

For file operations, Java will interpret the Unix file separator ("/") as if it were the file separator even when running in windows. Internally paths are used as keys and need to be consistent. Some of these paths are constructed programmatically, others are read from zip file directories, etc. It is better to adopt the simple standard to always use "/" as the separator.

Closes #26

PiperOrigin-RevId: 236383428