Fix load labels to respect package boundaries
In the label syntax, the position of the colon should match the package it belongs to (https://docs.bazel.build/versions/master/build-ref.html#labels). Due to a bug, Bazel was not checking that the position of the colon was correct; we are fixing this.
The check is enabled since Bazel 0.25.

PiperOrigin-RevId: 254427095