Remove meaningless test.
This tests claims to be safeguarding from having casttoclassdisabled optimizing to nothing (and not failing) not because of the cast optimizations, but for a incorrect optimization jscompiler (jscompiler removing side effectful code).  There is no value in this test. Also note that there is no size reduction and the size of this test is similar to casttoclass.

.

PiperOrigin-RevId: 270145460