Fix the implementation of getMethodsBySignature().
TypeDescriptor(TypeDeclaration).getMethodBySignature were including static methods and constructors from super classes incorrectly (instead of only inherited methods). The main reason why this did not emerge as a bug is that the usage is mostly to lookup standard runtime methods (like Emun.ordinal(), etc). There is only one usage in BridgeMethodsCreator involving lookup of methods in interfaces which could be a problem but it was not because the results were being filtered and excluded non polymorphic methods.

In summary, the fix should not have any effect in the current setup due to the limited usage of the API.

PiperOrigin-RevId: 270990118