Improve the warning when a native method is declared but is not a @JsMethod.
Native Java methods need to be provided as JavaScript code in separate files that have a ".native.js" suffix. If a native Java method is not a @JsMethod, name mangling occurs and such mangling is an internal implementation detail that is subject to change. Therefore, when a native method is not a @JsMethod, throw a specific warning.

PiperOrigin-RevId: 272457902