Make JS stack traces of J2CL exceptions a lot more precise and make the underlying JavaScript error message reflect Throwable.toString.
This will reduce down number extra frames to 1 frame which would be the Exception constructor (which is dropped in Java and pure JavaScript errors).

For browsers that use V8 (Chrome, Opera etc and soon IE) we could get down to no extra frames by using Error.captureStackTrace however that would add 2 frames to all other browsers and doesn't seem worthwhile at the moment. We will re-evaluate later.

For the error message part:
Throwable.toString() is sometimes overridden to provide more details about the exception (e.g. test infra). We are now including that in the backing JavaScript Error.
That couldn't be done before because calling toString before execution of the child constructor was dangerous.

PiperOrigin-RevId: 272791957