Emit the correct implicit super constructor call when the super is varargs.
Consider the following class:

  class A { A(String... args) { } }

when such class is extended like in

  class B extends A { B() { } }

but its constructor does not explicitly call the superconstructor, Java semantics imply that an implicit super() call, with no parameters) is inserted. That super() call is subject to overloads semantics and in this case the target of the call is the varargs constructor of A. So the super() call inserted is equivalent to super(new String[0]).

PiperOrigin-RevId: 273281793