Start tracking suppressed errors in addition to the cause in underlying error object.
With the CL the extra data is no longer tracked with quoted property so they can be removed when they are not used by app. The implication of that is, by default such properties won't be accessible outside of the binary. Such use case (if comes up) could be still handled via externs or via user utilities exposed from the binary.

PiperOrigin-RevId: 275914666