Model super calls that target a default method in the interface consistently.
Super method call always have as a target an existing method. Subsequent passes have always assumed that the method descriptor target of a super call in a super class. But with default methods, before this cl, that was not the case, since the target might be an interface default method.

PiperOrigin-RevId: 282022042