Add support for RTA in J2CL integration_tests.
We have a series of integration tests that verify the proper behavior of J2CL when RTA is disabled. We want to use those existing tests to also verify the proper behavior of RTA. Therefore, we modify the rule "integration_test" to provide the option of enabling RTA when running an integration_test.

PiperOrigin-RevId: 286442433