Stop tracking clinit since clinit is live if the type is referenced.
Only cases where the type is referenced but it might not be 'live' is the case where the type only shows up on things like instanceof.

We can improve our accuracy by recording the clinit record calls made from a 'clinit' instead of recording them as type reference. And then we can get away by considering clinit live iff there is a static call to type. However this is a minor small improvement in practice (<0.1%).

PiperOrigin-RevId: 293266813