Avoid unnecessary polyfill injection for async functions
Drop the require statements for Symbol and Promise polyfills from the runtime
libraries used to transpile async functions and generators.

These runtime libraries are now injected before rewritePolyfills runs,
so they will automatically trigger injection of the polyfills when they are needed.
Forced inclusion was causing the polyfills to remain in the output even when
not needed for the output language level.

PiperOrigin-RevId: 314761970