Use a principled approach to create bridge method signatures.
There are three separate components that are involved in the typing of the synthetic bridge methods:
  - the mangled name for the methods which has to be derived from the overridden method declaration.
  - the JsDoc typing for the method which is derived from the methods as parameterized by the class.
  - the actual type of the parameter variables which derived from the overridden declarations. This is because due to raw types the actual values might not satisfy the local typing, and runtime checks are added.

Bridge method creation relies in setting the method declaration to use the overridden method typing (used for mangling) and the actual method descriptor to use the locally parametrized typing.

In the past the typing for the actual parameter variables was used to declare the JsDoc types but it was incorrect when type variables related the return type to the type of a parameter.

PiperOrigin-RevId: 314838797