Fixes an incorrect optimization that removed ALL type references to a class if there were only references to an instance member of a class even if there were other type references (e.g. from type literals).
RTA preserves the enclosing type if there is a reference to one of his static members. But static member calls ALSO have a constructor reference as their qualifier, which is redundant and need not be preserved separately as a type reference.

The implementation fixed here removed type references even if they were only the enclosing class of an instance member which was incorrect.

PiperOrigin-RevId: 331825976