[WASM] Simplify loop rewriting.
By moving some of the loop emission logic to the output stage, is more compact and also the code is simpler here than as a rewriting pass. Some of the helpers introduced can be used by other constructs (e.g. switch).

Rewriting in the AST has the limitations of a source to source transformation making it more difficult to specialize the generation to make use of instructions like br_if that make the output more compact.

PiperOrigin-RevId: 350566707