Speed up transpiler by writing files asynchronously
Originally the output state is working like this:

while(moreInput) {
 do process output (computation)
 do bunch of I/O (ensure directories, write output, copy java files etc.)
}

That means we are not overlapping computation with I/O.

This wasn't a problem earlier because the I/O on zip files are effectively async since they are implemented like an in-memory file system. However when we output to a directory, we see a performance reqression.

This CL is making sure that computation overlaps with I/O by moving I/O to be done asynchronously on a different thread (but still in order). As a result instead of waiting for I/O to complete, we are now starting working on the next output.

PiperOrigin-RevId: 350670284