Add test for rta pruning a live type due to coalescing members in library info.
Library info coalesces members that have the same name together preserving the union of their references but only preserving the properties of one of them (whether it is static, their source info).
That is done to handle getters/setters as one member however this may also result in incorrect results when a static method has the same name as an instance method.

PiperOrigin-RevId: 351256636