Allow static native JsMethods to have an empty name
This would allow one to represent a function that is itself a namespace, but not
a type.

I was pretty strict with the restrictions on this. You need to explicitly
specify a namespace (on the JsMethod itself!) if you want to use this feature.
Inferring it from the enclosing descriptor gets weird, especially since it means
that you're representing the namespace as if it were a type.

PiperOrigin-RevId: 361227873