[WASM] Add Java type abstraction for Wasm arrays.
This allows us to have proper inheritance relation between Object and Array types. It also introduces a top level type for all arrays that enables generic array access that is needed for emulation of the some JRE APIs.

To make this work, a new type called javaemul.internal.WasmArray introduced with subtypes for all top level array types. In a very late pass, all arrays are rewritten to use these new types. After that only remaining real arrays are the ones under these classes which are special treated in the backend for code generation.

PiperOrigin-RevId: 365193143