Add support for native array operations needed by JRE.
One of the critical parts of the design is the introduction of length to support growing of the arrays. This behavior doesn't affect user arrays but enables growing arrays for JRE in order to reuse it from Closure version.

This patch doesn't add the operations to all array variations but adds the basic infrastructure in place for the general pattern. After proven, we may either copy it to other types or do some kind of templating to simplify the maintenance.

PiperOrigin-RevId: 368952505