[WASM] Implement interface dispatch.
Each object has a field $itable which is an array of interface vtables. These interface vtables only contain the methods declared in the corresponding interface.
A particular interface if implemented will inhabit always the same slot in the itable, so the same slot can be shared by multiple interfaces. Interfaces that share a slot must be implemented by disjoint sets of concrete classes.

Slots are allocated greedily, trying to assign lower numbers to interfaces with larger number of concrete implementors.

Interface dispatch consists of accessing the vtable slot corresponding to the interface in question, casting the vtable to the corresponding interface vtable type and dispatching to the corresponding vtable function.

PiperOrigin-RevId: 370582301